+++ src/snowflake/snowpark/files.py
@@ -79,7 +79,9 @@ class SnowflakeFile(RawIOBase):
 
     These examples are using the client, but this same pattern can be used inside SQL-defined UDFs.
 
-    We provide a local implementation of SnowflakeFile to aid in local testing. This currently only supports using read APIs on relative paths and mocked stages.
+    We provide a local implementation of SnowflakeFile to aid in local testing.
+    This currently only supports using read APIs on relative paths and mocked stages
+    (sessions in local testing mode that aren't connected to a real stage).
 
     Note:
         1. All of the implementation in this file is for local testing purposes.